Module google_content2_sandbox::api

Structs

Amount
There is no detailed description.

CustomerReturnReason
There is no detailed description.

Error
An error returned by the API.

Errors
A list of errors returned by a failed batch entry.

InvoiceSummary
There is no detailed description.

InvoiceSummaryAdditionalChargeSummary
There is no detailed description.

MerchantOrderReturn
There is no detailed description.

MerchantOrderReturnItem
There is no detailed description.

Order
There is no detailed description.

OrderAcknowledgeCall
Marks an order as acknowledged.

OrderAddress
There is no detailed description.

OrderAdvancetestorderCall
Sandbox only. Moves a test order from state "inProgress" to state "pendingShipment".

OrderCancelCall
Cancels all line items in an order, making a full refund.

OrderCancellation
There is no detailed description.

OrderCancellineitemCall
Cancels a line item, making a full refund.

OrderCanceltestorderbycustomerCall
Sandbox only. Cancels a test order for customer-initiated cancellation.

OrderCreatetestorderCall
Sandbox only. Creates a test order.

OrderCreatetestreturnCall
Sandbox only. Creates a test return.

OrderCustombatchCall
Retrieves or modifies multiple orders in a single request.

OrderCustomer
There is no detailed description.

OrderCustomerMarketingRightsInfo
There is no detailed description.

OrderDeliveryDetails
There is no detailed description.

OrderGetCall
Retrieves an order from your Merchant Center account.

OrderGetbymerchantorderidCall
Retrieves an order using merchant order id.

OrderGettestordertemplateCall
Sandbox only. Retrieves an order template that can be used to quickly create a new order in sandbox.

OrderInstorerefundlineitemCall
Notifies that item return and refund was handled directly by merchant outside of Google payments processing (e.g. cash refund done in store).

OrderLegacyPromotion
There is no detailed description.

OrderLegacyPromotionBenefit
There is no detailed description.

OrderLineItem
There is no detailed description.

OrderLineItemProduct
There is no detailed description.

OrderLineItemProductVariantAttribute
There is no detailed description.

OrderLineItemReturnInfo
There is no detailed description.

OrderLineItemShippingDetails
There is no detailed description.

OrderLineItemShippingDetailsMethod
There is no detailed description.

OrderListCall
Lists the orders in your Merchant Center account.

OrderMerchantProvidedAnnotation
There is no detailed description.

OrderMethods
A builder providing access to all methods supported on order resources. It is not used directly, but through the ShoppingContent hub.

OrderPaymentMethod
There is no detailed description.

OrderRefund
There is no detailed description.

OrderRefundCall
Deprecated, please use returnRefundLineItem instead.

OrderRejectreturnlineitemCall
Rejects return on an line item.

OrderReturn
There is no detailed description.

OrderReturnlineitemCall
Returns a line item.

OrderReturnrefundlineitemCall
Returns and refunds a line item. Note that this method can only be called on fully shipped orders.

OrderSetlineitemmetadataCall
Sets (overrides) merchant provided annotations on the line item.

OrderShiplineitemCall
Marks line item(s) as shipped.

OrderShipment
There is no detailed description.

OrderShipmentLineItemShipment
There is no detailed description.

OrderUpdatelineitemshippingdetailCall
Updates ship by and delivery by dates for a line item.

OrderUpdatemerchantorderidCall
Updates the merchant order ID for a given order.

OrderUpdateshipmentCall
Updates a shipment's status, carrier, and/or tracking ID.

OrderinvoiceCreatechargeinvoiceCall
Creates a charge invoice for a shipment group, and triggers a charge capture for non-facilitated payment orders.

OrderinvoiceCreaterefundinvoiceCall
Creates a refund invoice for one or more shipment groups, and triggers a refund for non-facilitated payment orders. This can only be used for line items that have previously been charged using createChargeInvoice. All amounts (except for the summary) are incremental with respect to the previous invoice.

OrderinvoiceMethods
A builder providing access to all methods supported on orderinvoice resources. It is not used directly, but through the ShoppingContent hub.

OrderinvoicesCreateChargeInvoiceRequest
There is no detailed description.

OrderinvoicesCreateChargeInvoiceResponse
There is no detailed description.

OrderinvoicesCreateRefundInvoiceRequest
There is no detailed description.

OrderinvoicesCreateRefundInvoiceResponse
There is no detailed description.

OrderinvoicesCustomBatchRequestEntryCreateRefundInvoiceRefundOption
There is no detailed description.

OrderinvoicesCustomBatchRequestEntryCreateRefundInvoiceReturnOption
There is no detailed description.

OrderpaymentMethods
A builder providing access to all methods supported on orderpayment resources. It is not used directly, but through the ShoppingContent hub.

OrderpaymentNotifyauthapprovedCall
Notify about successfully authorizing user's payment method for a given amount.

OrderpaymentNotifyauthdeclinedCall
Notify about failure to authorize user's payment method.

OrderpaymentNotifychargeCall
Notify about charge on user's selected payments method.

OrderpaymentNotifyrefundCall
Notify about refund on user's selected payments method.

OrderpaymentsNotifyAuthApprovedRequest
There is no detailed description.

OrderpaymentsNotifyAuthApprovedResponse
There is no detailed description.

OrderpaymentsNotifyAuthDeclinedRequest
There is no detailed description.

OrderpaymentsNotifyAuthDeclinedResponse
There is no detailed description.

OrderpaymentsNotifyChargeRequest
There is no detailed description.

OrderpaymentsNotifyChargeResponse
There is no detailed description.

OrderpaymentsNotifyRefundRequest
There is no detailed description.

OrderpaymentsNotifyRefundResponse
There is no detailed description.

OrderreturnGetCall
Retrieves an order return from your Merchant Center account.

OrderreturnListCall
Lists order returns in your Merchant Center account.

OrderreturnMethods
A builder providing access to all methods supported on orderreturn resources. It is not used directly, but through the ShoppingContent hub.

OrderreturnsListResponse
There is no detailed description.

OrdersAcknowledgeRequest
There is no detailed description.

OrdersAcknowledgeResponse
There is no detailed description.

OrdersAdvanceTestOrderResponse
There is no detailed description.

OrdersCancelLineItemRequest
There is no detailed description.

OrdersCancelLineItemResponse
There is no detailed description.

OrdersCancelRequest
There is no detailed description.

OrdersCancelResponse
There is no detailed description.

OrdersCancelTestOrderByCustomerRequest
There is no detailed description.

OrdersCancelTestOrderByCustomerResponse
There is no detailed description.

OrdersCreateTestOrderRequest
There is no detailed description.

OrdersCreateTestOrderResponse
There is no detailed description.

OrdersCreateTestReturnRequest
There is no detailed description.

OrdersCreateTestReturnResponse
There is no detailed description.

OrdersCustomBatchRequest
There is no detailed description.

OrdersCustomBatchRequestEntry
There is no detailed description.

OrdersCustomBatchRequestEntryCancel
There is no detailed description.

OrdersCustomBatchRequestEntryCancelLineItem
There is no detailed description.

OrdersCustomBatchRequestEntryCreateTestReturnReturnItem
There is no detailed description.

OrdersCustomBatchRequestEntryInStoreRefundLineItem
There is no detailed description.

OrdersCustomBatchRequestEntryRefund
There is no detailed description.

OrdersCustomBatchRequestEntryRejectReturnLineItem
There is no detailed description.

OrdersCustomBatchRequestEntryReturnLineItem
There is no detailed description.

OrdersCustomBatchRequestEntryReturnRefundLineItem
There is no detailed description.

OrdersCustomBatchRequestEntrySetLineItemMetadata
There is no detailed description.

OrdersCustomBatchRequestEntryShipLineItems
There is no detailed description.

OrdersCustomBatchRequestEntryShipLineItemsShipmentInfo
There is no detailed description.

OrdersCustomBatchRequestEntryUpdateLineItemShippingDetails
There is no detailed description.

OrdersCustomBatchRequestEntryUpdateShipment
There is no detailed description.

OrdersCustomBatchResponse
There is no detailed description.

OrdersCustomBatchResponseEntry
There is no detailed description.

OrdersGetByMerchantOrderIdResponse
There is no detailed description.

OrdersGetTestOrderTemplateResponse
There is no detailed description.

OrdersInStoreRefundLineItemRequest
There is no detailed description.

OrdersInStoreRefundLineItemResponse
There is no detailed description.

OrdersListResponse
There is no detailed description.

OrdersRefundRequest
There is no detailed description.

OrdersRefundResponse
There is no detailed description.

OrdersRejectReturnLineItemRequest
There is no detailed description.

OrdersRejectReturnLineItemResponse
There is no detailed description.

OrdersReturnLineItemRequest
There is no detailed description.

OrdersReturnLineItemResponse
There is no detailed description.

OrdersReturnRefundLineItemRequest
There is no detailed description.

OrdersReturnRefundLineItemResponse
There is no detailed description.

OrdersSetLineItemMetadataRequest
There is no detailed description.

OrdersSetLineItemMetadataResponse
There is no detailed description.

OrdersShipLineItemsRequest
There is no detailed description.

OrdersShipLineItemsResponse
There is no detailed description.

OrdersUpdateLineItemShippingDetailsRequest
There is no detailed description.

OrdersUpdateLineItemShippingDetailsResponse
There is no detailed description.

OrdersUpdateMerchantOrderIdRequest
There is no detailed description.

OrdersUpdateMerchantOrderIdResponse
There is no detailed description.

OrdersUpdateShipmentRequest
There is no detailed description.

OrdersUpdateShipmentResponse
There is no detailed description.

Price
There is no detailed description.

Promotion
There is no detailed description.

RefundReason
There is no detailed description.

ReturnShipment
There is no detailed description.

ShipmentInvoice
There is no detailed description.

ShipmentInvoiceLineItemInvoice
There is no detailed description.

ShipmentTrackingInfo
There is no detailed description.

ShoppingContent
Central instance to access all ShoppingContent related resource activities

TestOrder
There is no detailed description.

TestOrderCustomer
There is no detailed description.

TestOrderCustomerMarketingRightsInfo
There is no detailed description.

TestOrderLineItem
There is no detailed description.

TestOrderLineItemProduct
There is no detailed description.

TestOrderPaymentMethod
There is no detailed description.

UnitInvoice
There is no detailed description.

UnitInvoiceAdditionalCharge
There is no detailed description.

UnitInvoiceTaxLine
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.