Struct google_content2_sandbox::OrderLineItem

pub struct OrderLineItem {
    pub product: Option<OrderLineItemProduct>,
    pub quantity_delivered: Option<u32>,
    pub return_info: Option<OrderLineItemReturnInfo>,
    pub price: Option<Price>,
    pub quantity_pending: Option<u32>,
    pub tax: Option<Price>,
    pub cancellations: Option<Vec<OrderCancellation>>,
    pub quantity_canceled: Option<u32>,
    pub id: Option<String>,
    pub returns: Option<Vec<OrderReturn>>,
    pub quantity_shipped: Option<u32>,
    pub quantity_returned: Option<u32>,
    pub quantity_ordered: Option<u32>,
    pub shipping_details: Option<OrderLineItemShippingDetails>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
product: Option<OrderLineItemProduct>

Product data from the time of the order placement.

quantity_delivered: Option<u32>

Number of items delivered.

return_info: Option<OrderLineItemReturnInfo>

Details of the return policy for the line item.

price: Option<Price>

Total price for the line item. For example, if two items for $10 are purchased, the total price will be $20.

quantity_pending: Option<u32>

Number of items pending.

tax: Option<Price>

Total tax amount for the line item. For example, if two items are purchased, and each have a cost tax of $2, the total tax amount will be $4.

cancellations: Option<Vec<OrderCancellation>>

Cancellations of the line item.

quantity_canceled: Option<u32>

Number of items canceled.

id: Option<String>

The id of the line item.

returns: Option<Vec<OrderReturn>>

Returns of the line item.

quantity_shipped: Option<u32>

Number of items shipped.

quantity_returned: Option<u32>

Number of items returned.

quantity_ordered: Option<u32>

Number of items ordered.

shipping_details: Option<OrderLineItemShippingDetails>

Details of the requested shipping for the line item.