Struct google_content2_sandbox::OrderMethods

pub struct OrderMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on order resources. It is not used directly, but through the ShoppingContent hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_content2_sandbox as content2_sandbox;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use content2_sandbox::ShoppingContent;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = ShoppingContent::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `acknowledge(...)`, `advancetestorder(...)`, `cancel(...)`, `cancellineitem(...)`, `canceltestorderbycustomer(...)`, `createtestorder(...)`, `createtestreturn(...)`, `custombatch(...)`, `get(...)`, `getbymerchantorderid(...)`, `gettestordertemplate(...)`, `instorerefundlineitem(...)`, `list(...)`, `refund(...)`, `rejectreturnlineitem(...)`, `returnlineitem(...)`, `returnrefundlineitem(...)`, `setlineitemmetadata(...)`, `shiplineitems(...)`, `updatelineitemshippingdetails(...)`, `updatemerchantorderid(...)` and `updateshipment(...)`
// to build up your call.
let rb = hub.orders();

Methods

impl<'a, C, A> OrderMethods<'a, C, A>

pub fn list(&self, merchant_id: &str) -> OrderListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the orders in your Merchant Center account.

Arguments

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

pub fn updatemerchantorderid(
&self,
request: OrdersUpdateMerchantOrderIdRequest,
merchant_id: &str,
order_id: &str
) -> OrderUpdatemerchantorderidCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the merchant order ID for a given order.

Arguments

request - No description provided.

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the order.

pub fn gettestordertemplate(
&self,
merchant_id: &str,
template_name: &str
) -> OrderGettestordertemplateCall<'a, C, A>

Create a builder to help you perform the following task:

Sandbox only. Retrieves an order template that can be used to quickly create a new order in sandbox.

Arguments

merchantId - The ID of the account that should manage the order. This cannot be a multi-client account.

templateName - The name of the template to retrieve.

pub fn refund(
&self,
request: OrdersRefundRequest,
merchant_id: &str,
order_id: &str
) -> OrderRefundCall<'a, C, A>

Create a builder to help you perform the following task:

Deprecated, please use returnRefundLineItem instead.

Arguments

request - No description provided.

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the order to refund.

pub fn acknowledge(
&self,
request: OrdersAcknowledgeRequest,
merchant_id: &str,
order_id: &str
) -> OrderAcknowledgeCall<'a, C, A>

Create a builder to help you perform the following task:

Marks an order as acknowledged.

Arguments

request - No description provided.

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the order.

pub fn advancetestorder(
&self,
merchant_id: &str,
order_id: &str
) -> OrderAdvancetestorderCall<'a, C, A>

Create a builder to help you perform the following task:

Sandbox only. Moves a test order from state "inProgress" to state "pendingShipment".

Arguments

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the test order to modify.

pub fn rejectreturnlineitem(
&self,
request: OrdersRejectReturnLineItemRequest,
merchant_id: &str,
order_id: &str
) -> OrderRejectreturnlineitemCall<'a, C, A>

Create a builder to help you perform the following task:

Rejects return on an line item.

Arguments

request - No description provided.

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the order.

pub fn createtestorder(
&self,
request: OrdersCreateTestOrderRequest,
merchant_id: &str
) -> OrderCreatetestorderCall<'a, C, A>

Create a builder to help you perform the following task:

Sandbox only. Creates a test order.

Arguments

request - No description provided.

merchantId - The ID of the account that should manage the order. This cannot be a multi-client account.

pub fn shiplineitems(
&self,
request: OrdersShipLineItemsRequest,
merchant_id: &str,
order_id: &str
) -> OrderShiplineitemCall<'a, C, A>

Create a builder to help you perform the following task:

Marks line item(s) as shipped.

Arguments

request - No description provided.

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the order.

pub fn canceltestorderbycustomer(
&self,
request: OrdersCancelTestOrderByCustomerRequest,
merchant_id: &str,
order_id: &str
) -> OrderCanceltestorderbycustomerCall<'a, C, A>

Create a builder to help you perform the following task:

Sandbox only. Cancels a test order for customer-initiated cancellation.

Arguments

request - No description provided.

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the test order to cancel.

pub fn returnrefundlineitem(
&self,
request: OrdersReturnRefundLineItemRequest,
merchant_id: &str,
order_id: &str
) -> OrderReturnrefundlineitemCall<'a, C, A>

Create a builder to help you perform the following task:

Returns and refunds a line item. Note that this method can only be called on fully shipped orders.

Arguments

request - No description provided.

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the order.

pub fn get(&self, merchant_id: &str, order_id: &str) -> OrderGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves an order from your Merchant Center account.

Arguments

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the order.

pub fn returnlineitem(
&self,
request: OrdersReturnLineItemRequest,
merchant_id: &str,
order_id: &str
) -> OrderReturnlineitemCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a line item.

Arguments

request - No description provided.

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the order.

pub fn updatelineitemshippingdetails(
&self,
request: OrdersUpdateLineItemShippingDetailsRequest,
merchant_id: &str,
order_id: &str
) -> OrderUpdatelineitemshippingdetailCall<'a, C, A>

Create a builder to help you perform the following task:

Updates ship by and delivery by dates for a line item.

Arguments

request - No description provided.

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the order.

pub fn custombatch(
&self,
request: OrdersCustomBatchRequest
) -> OrderCustombatchCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves or modifies multiple orders in a single request.

Arguments

request - No description provided.

pub fn instorerefundlineitem(
&self,
request: OrdersInStoreRefundLineItemRequest,
merchant_id: &str,
order_id: &str
) -> OrderInstorerefundlineitemCall<'a, C, A>

Create a builder to help you perform the following task:

Notifies that item return and refund was handled directly by merchant outside of Google payments processing (e.g. cash refund done in store).

Arguments

request - No description provided.

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the order.

pub fn cancellineitem(
&self,
request: OrdersCancelLineItemRequest,
merchant_id: &str,
order_id: &str
) -> OrderCancellineitemCall<'a, C, A>

Create a builder to help you perform the following task:

Cancels a line item, making a full refund.

Arguments

request - No description provided.

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the order.

pub fn createtestreturn(
&self,
request: OrdersCreateTestReturnRequest,
merchant_id: &str,
order_id: &str
) -> OrderCreatetestreturnCall<'a, C, A>

Create a builder to help you perform the following task:

Sandbox only. Creates a test return.

Arguments

request - No description provided.

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the order.

pub fn updateshipment(
&self,
request: OrdersUpdateShipmentRequest,
merchant_id: &str,
order_id: &str
) -> OrderUpdateshipmentCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a shipment's status, carrier, and/or tracking ID.

Arguments

request - No description provided.

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the order.

pub fn getbymerchantorderid(
&self,
merchant_id: &str,
merchant_order_id: &str
) -> OrderGetbymerchantorderidCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves an order using merchant order id.

Arguments

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

merchantOrderId - The merchant order id to be looked for.

pub fn setlineitemmetadata(
&self,
request: OrdersSetLineItemMetadataRequest,
merchant_id: &str,
order_id: &str
) -> OrderSetlineitemmetadataCall<'a, C, A>

Create a builder to help you perform the following task:

Sets (overrides) merchant provided annotations on the line item.

Arguments

request - No description provided.

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the order.

pub fn cancel(
&self,
request: OrdersCancelRequest,
merchant_id: &str,
order_id: &str
) -> OrderCancelCall<'a, C, A>

Create a builder to help you perform the following task:

Cancels all line items in an order, making a full refund.

Arguments

request - No description provided.

merchantId - The ID of the account that manages the order. This cannot be a multi-client account.

orderId - The ID of the order to cancel.