Struct google_content2_sandbox::OrderMethods

pub struct OrderMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on order resources. It is not used directly, but through the ShoppingContent hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_content2_sandbox as content2_sandbox;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use content2_sandbox::ShoppingContent;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = ShoppingContent::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `acknowledge(...)`, `advancetestorder(...)`, `cancel(...)`, `cancellineitem(...)`, `createtestorder(...)`, `custombatch(...)`, `get(...)`, `getbymerchantorderid(...)`, `gettestordertemplate(...)`, `list(...)`, `refund(...)`, `returnlineitem(...)`, `shiplineitems(...)`, `updatemerchantorderid(...)` and `updateshipment(...)`
// to build up your call.
let rb = hub.orders();

Methods

impl<'a, C, A> OrderMethods<'a, C, A>

fn get(&self, merchant_id: &str, order_id: &str) -> OrderGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves an order from your Merchant Center account.

Arguments

merchantId - The ID of the managing account.

orderId - The ID of the order.

fn list(&self, merchant_id: &str) -> OrderListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the orders in your Merchant Center account.

Arguments

merchantId - The ID of the managing account.

fn updateshipment(&self, request: OrdersUpdateShipmentRequest, merchant_id: &str, order_id: &str) -> OrderUpdateshipmentCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a shipment's status, carrier, and/or tracking ID.

Arguments

request - No description provided.

merchantId - The ID of the managing account.

orderId - The ID of the order.

fn advancetestorder(&self, merchant_id: &str, order_id: &str) -> OrderAdvancetestorderCall<'a, C, A>

Create a builder to help you perform the following task:

Sandbox only. Moves a test order from state "inProgress" to state "pendingShipment".

Arguments

merchantId - The ID of the managing account.

orderId - The ID of the test order to modify.

fn updatemerchantorderid(&self, request: OrdersUpdateMerchantOrderIdRequest, merchant_id: &str, order_id: &str) -> OrderUpdatemerchantorderidCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the merchant order ID for a given order.

Arguments

request - No description provided.

merchantId - The ID of the managing account.

orderId - The ID of the order.

fn returnlineitem(&self, request: OrdersReturnLineItemRequest, merchant_id: &str, order_id: &str) -> OrderReturnlineitemCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a line item.

Arguments

request - No description provided.

merchantId - The ID of the managing account.

orderId - The ID of the order.

fn gettestordertemplate(&self, merchant_id: &str, template_name: &str) -> OrderGettestordertemplateCall<'a, C, A>

Create a builder to help you perform the following task:

Sandbox only. Retrieves an order template that can be used to quickly create a new order in sandbox.

Arguments

merchantId - The ID of the managing account.

templateName - The name of the template to retrieve.

fn createtestorder(&self, request: OrdersCreateTestOrderRequest, merchant_id: &str) -> OrderCreatetestorderCall<'a, C, A>

Create a builder to help you perform the following task:

Sandbox only. Creates a test order.

Arguments

request - No description provided.

merchantId - The ID of the managing account.

fn refund(&self, request: OrdersRefundRequest, merchant_id: &str, order_id: &str) -> OrderRefundCall<'a, C, A>

Create a builder to help you perform the following task:

Refund a portion of the order, up to the full amount paid.

Arguments

request - No description provided.

merchantId - The ID of the managing account.

orderId - The ID of the order to refund.

fn custombatch(&self, request: OrdersCustomBatchRequest) -> OrderCustombatchCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves or modifies multiple orders in a single request.

Arguments

request - No description provided.

fn cancellineitem(&self, request: OrdersCancelLineItemRequest, merchant_id: &str, order_id: &str) -> OrderCancellineitemCall<'a, C, A>

Create a builder to help you perform the following task:

Cancels a line item.

Arguments

request - No description provided.

merchantId - The ID of the managing account.

orderId - The ID of the order.

fn getbymerchantorderid(&self, merchant_id: &str, merchant_order_id: &str) -> OrderGetbymerchantorderidCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves an order using merchant order id.

Arguments

merchantId - The ID of the managing account.

merchantOrderId - The merchant order id to be looked for.

fn acknowledge(&self, request: OrdersAcknowledgeRequest, merchant_id: &str, order_id: &str) -> OrderAcknowledgeCall<'a, C, A>

Create a builder to help you perform the following task:

Marks an order as acknowledged.

Arguments

request - No description provided.

merchantId - The ID of the managing account.

orderId - The ID of the order.

fn cancel(&self, request: OrdersCancelRequest, merchant_id: &str, order_id: &str) -> OrderCancelCall<'a, C, A>

Create a builder to help you perform the following task:

Cancels all line items in an order.

Arguments

request - No description provided.

merchantId - The ID of the managing account.

orderId - The ID of the order to cancel.

fn shiplineitems(&self, request: OrdersShipLineItemsRequest, merchant_id: &str, order_id: &str) -> OrderShiplineitemCall<'a, C, A>

Create a builder to help you perform the following task:

Marks line item(s) as shipped.

Arguments

request - No description provided.

merchantId - The ID of the managing account.

orderId - The ID of the order.